server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-pptx==0.6.21
pytesseract==0.3.10
Pillow==10.0.1
openai==1.3.5
google-cloud-texttospeech==2.16.3
requests==2.31.0
pydub==0.25.1
ffmpeg-python==0.2.0
